Move character-counter-directive out of its own directory.
In Google, every directory containing TypeScript files must be a
directory with a BUILD file (a package), and packages cannot contain
dashes. We could handle this case with a transformation, but it seems
simpler to just keep a flatter layout upstream.
