+++ packages/textfield/src/mwc-character-counter-directive.ts
